Backport 697f27c5d53dbe275685b87c8ed1bcfe4da6e4d0
